% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_ludis_datasets}
\alias{list_ludis_datasets}
\title{Add together two numbers}
\usage{
list_ludis_datasets(url, headers)
}
\arguments{
\item{url}{url corresponding to datasets s3 location}

\item{headers}{headers containing a ludis api token}
}
\value{
a dataframe \code{df}
}
\description{
Add together two numbers
}
