% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_ludis_csv}
\alias{read_ludis_csv}
\title{Add together two numbers}
\usage{
read_ludis_csv(url, headers)
}
\arguments{
\item{url}{url corresponding to datasets s3 location}

\item{headers}{headers containing a ludis api token}
}
\value{
a dataframe \code{df}
}
\description{
Add together two numbers
}
