% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_ludis_csv}
\alias{write_ludis_csv}
\title{Add together two numbers}
\usage{
write_ludis_csv(url, headers)
}
\arguments{
\item{url}{url corresponding to datasets s3 location}

\item{headers}{headers containing a ludis api tokenfdsfdsa}
}
\value{
return a httr::PUT reponse object
}
\description{
Add together two numbers
}
